rename tablesynchronizer + typescomparator
